refactor: move api and components folders to js folder
