+++ src/js/app.js
@@ -1,5 +1,5 @@
 import '../scss/main.scss';
-import {Finder} from '../components/finder/finder';
+import {Finder} from './components/finder/finder';
 
 const header = document.getElementById('header');
 const searchForm = new Finder(header);
