+++ src/js/components/slider/slider.js
