+++ src/js/components/slider/slider_cards.js
