+++ src/js/components/slider/slider_cards.scss
@@ -1,4 +1,4 @@
-@import "../../scss/utils/vars";
+@import "../../../scss/utils/vars";
 
 .video {
   display: flex;
