+++ src/js/components/slider/slider_page.scss
