+++ src/js/components/slider/slider_paging-control.js
