+++ src/js/components/spinner/spinner.js
