+++ src/js/components/spinner/spinner.scss
