+++ src/js/components/video-manager/video-manager.js
