+++ src/scss/main.scss
@@ -4,9 +4,9 @@
 @import "common/label";
 @import "common/common";
 
-@import "../components/finder/finder";
-@import "../components/slider/slider";
-@import "../components/slider/slider_page";
-@import "../components/slider/slider_cards";
-@import "../components/slider/slider_paging-control";
-@import "../components/spinner/spinner";
+@import "../js/components/finder/finder";
+@import "../js/components/slider/slider";
+@import "../js/components/slider/slider_page";
+@import "../js/components/slider/slider_cards";
+@import "../js/components/slider/slider_paging-control";
+@import "../js/components/spinner/spinner";
